Move the tests back and skip
